rename: components/auth에 depth추가
