+++ src/pages/signin/index.page.tsx
@@ -1,6 +1,6 @@
 import router from "next/router";
 
-import SigninForm from "@/components/auth/SigninForm";
+import SigninForm from "@/components/auth/SigninForm/SigninForm";
 import useToast from "@/hooks/useToast";
 import getLatestYearMonthString from "@/utils/getLatestYearMonthString";
